Frontend/requirements.txt
streamlit
requests
pandas
snowflake-connector-python
